requirements/base.txt
Django==1.8.7
django-tinymce==2.0.6
Pillow==3.0.0
psycopg2==2.6.1
wheel==0.24.0
